test: fix test import paths and move test-utils to correct location
